Need to fix ouputFormat
